+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".ArtistListActivity"
+            android:name=".MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -18,15 +18,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".ArtistDetailActivity"
-            android:label="@string/title_artist_detail"
-            android:parentActivityName=".ArtistListActivity"
-            android:theme="@style/AppTheme.NoActionBar">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.example.grin.mobilizationmusic.ArtistListActivity" />
-        </activity>
+
         <service
             android:name="com.example.grin.mobilizationmusic.authentication.AuthenticatorService">
             <intent-filter>
