+++ app/src/main/java/com/example/grin/mobilizationmusic/ArtistDetailActivity.java
@@ -2,10 +2,7 @@ package com.example.grin.mobilizationmusic;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.widget.Toolbar;
-import android.view.View;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.app.ActionBar;
 import android.view.MenuItem;
@@ -14,7 +11,7 @@ import android.view.MenuItem;
  * An activity representing a single Artist detail screen. This
  * activity is only used narrow width devices. On tablet-size devices,
  * item details are presented side-by-side with a list of items
- * in a {@link ArtistListActivity}.
+ * in a {@link MainActivity}.
  */
 public class ArtistDetailActivity extends AppCompatActivity {
     @Override
@@ -66,7 +63,7 @@ public class ArtistDetailActivity extends AppCompatActivity {
             //
             // http://developer.android.com/design/patterns/navigation.html#up-vs-back
             //
-            navigateUpTo(new Intent(this, ArtistListActivity.class));
+            navigateUpTo(new Intent(this, MainActivity.class));
             return true;
         }
         return super.onOptionsItemSelected(item);
