+++ app/src/main/java/com/example/grin/mobilizationmusic/ArtistListFragment.java
@@ -1,0 +1,17 @@
+package com.example.grin.mobilizationmusic;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+/**
+ * Created by grin3s on 20.07.16.
+ */
+public class ArtistListFragment extends Fragment {
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.artist_detail, container, false);
+    }
+}
