+++ app/src/main/res/layout-w900dp/artist_list.xml
@@ -9,7 +9,7 @@
     android:divider="?android:attr/dividerHorizontal"
     android:orientation="horizontal"
     android:showDividers="middle"
-    tools:context="com.example.grin.mobilizationmusic.ArtistListActivity">
+    tools:context=".MainActivity">
 
     <!--
     This layout is a two-pane layout for the Artists
@@ -25,7 +25,7 @@
         android:layout_height="match_parent"
         android:layout_marginLeft="16dp"
         android:layout_marginRight="16dp"
-        tools:context="com.example.grin.mobilizationmusic.ArtistListActivity"
+        tools:context=".MainActivity"
         tools:listitem="@layout/artist_list_content" />
 
     <FrameLayout
